Datacamp - Career Track Data Science
Project for the Data science career Track on Datacamp.

The business objective is to devise strategies to drive growth and retention.

Project consists of the analysis of the Android app market by comparing over ten thousand apps in Google Play across different categories.

The steps are classical Data loading, transformation and exploration in order to find relationships between app prices on Google app.